requirements.txt
streamlit>=1.30
Pillow>=9.0
requests>=2.28
# Optional (only if you want local BLIP captioning):
# transformers>=4.30
# torch
